Bots/Facebook/requirements.txt
Flask>=3.0
python-dotenv>=1.0
requests>=2.32
gunicorn>=21.2